Fix - remove EDIT button
Locally, the fix in the plugin to hide the EDIT button works, but on GitHub, the action would have installed the public version of the plugin without our fix, and the EDIT button would have been visible.

I will send a pull request to the upstream project of the plugin. For now, as a workaround, I have transformed the plugin into a hook